+++ Tests/RetentionFixtures/testIgnoreComments.swift
@@ -7,6 +7,11 @@ public class Fixture113 {
     }
 }
 
+public protocol Fixture114Protocol {
+    // periphery:ignore:parameters param
+    func protocolFunc(param: String)
+}
+
 public class Fixture114 {
     func referencedFunc() {}
 
@@ -14,4 +19,9 @@ public class Fixture114 {
     public func someFunc(a: String, b: String, c: String) {
         print(a)
     }
+}
+
+extension Fixture114: Fixture114Protocol {
+    // param is ignored becuse the protocol ignores it.
+    public func protocolFunc(param: String) {}
 }